refactor(Folders):
- Create in a separate files, spy and sut
