+++ SingletonAndGlobalInstances/SingletonAndGlobalInstancesTests/Scenes/Login/LoginClientAdapterTests.swift
@@ -27,26 +27,3 @@ final class LoginClientAdapterTests: XCTestCase {
         XCTAssertNotNil(expectedDataObject)
     }
 }
-
-final class LoginClientAdapterSpy: ApiClient {
-    var urlRequest: URL?
-    var expected: (Data)?
-    
-    private(set) var executeCalled: Bool = false
-    
-    override func execute(request: URLRequest, completion: (Data) -> Void) {
-        executeCalled = true
-        urlRequest = request.url
-        if let expected {
-            completion(expected)
-        }
-    }
-}
-
-extension LoginClientAdapterTests {
-    private func makeSut() -> (sut: LoginClientAdapter, adapterSpy: LoginClientAdapterSpy) {
-        let adapterSpy = LoginClientAdapterSpy()
-        let sut = LoginClientAdapter(api: adapterSpy)
-        return (sut, adapterSpy)
-    }
-}
